change component candidadate status
