+++ src/sesame/Vacancy/Infrastructure/Vue/components/VacancyBoard/OVacancyBoard.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import type { VacancyCandidateStatusDTO } from '@/sesame/Vacancy/Application/get-vacancy-candidate-status/VacancyCandidateStatusDTO'
 
-import OVacancyCandidateCard from './OVacancyCandidateCard.vue'
+import OVacancyCandidateStatusCard from './OVacancyCandidateStatusCard.vue'
 
 interface Props {
   candidateStatuses: VacancyCandidateStatusDTO[]
@@ -12,7 +12,7 @@ const { candidateStatuses } = defineProps<Props>()
 
 <template>
   <div class="flex items-stretch gap-3 whitespace-nowrap overflow-x-scroll">
-    <OVacancyCandidateCard
+    <OVacancyCandidateStatusCard
       v-for="candidateStatus in candidateStatuses"
       :key="candidateStatus.id"
       :candidate-status="candidateStatus"
